requirements.txt
aliyun-python-sdk-core==2.13.26
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.12.0
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
crcmod==1.7
decorator==4.4.2
Django==2.2.12
django-celery-beat==1.4.0
django-cors-headers==3.5.0
django-timezone-field==4.0
djangorestframework==3.11.1
dnspython==1.16.0
eventlet==0.25.2
flower==0.9.4
greenlet==0.4.17
humanize==0.5.1
idna==2.10
importlib-metadata==2.0.0
jmespath==0.10.0
kombu==4.6.11
monotonic==1.5
mysqlclient==2.0.1
oss2==2.13.0
prompt-toolkit==3.0.7
py==1.9.0
pycryptodome==3.9.8
pyDes==2.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
retry==0.9.2
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
zipp==3.2.0
